% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{format_naira}
\alias{format_naira}
\title{Format a number as Nigerian Naira}
\usage{
format_naira(x, prefix = "₦", digits = 2)
}
\arguments{
\item{x}{Numeric value.}

\item{prefix}{Character. Currency prefix. Default \code{"\u20a6"}.}

\item{digits}{Integer. Decimal places. Default \code{2}.}
}
\value{
A formatted character string.
}
\description{
Format a number as Nigerian Naira
}
\examples{
format_naira(1624734)
}
