% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/law_registry.R
\docType{package}
\name{nrsR-package}
\alias{nrsR}
\alias{nrsR-package}
\title{nrsr: Nigeria Revenue Service R Toolkit}
\description{
PAYE tax computation for Nigeria, driven by a versioned law registry.
Named after the Nigeria Revenue Service (NRS), which replaced FIRS
under the 2025 tax reforms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/laws2020/nrsr}
  \item Report bugs at \url{https://github.com/laws2020/nrsr/issues}
}

}
\author{
\strong{Maintainer}: Lawrence Garba \email{laws.garba@gmail.com} (Infosights Consulting)

Authors:
\itemize{
  \item Lawrence Garba \email{laws.garba@gmail.com} (Infosights Consulting)
}

}
\keyword{internal}
